serverless_app_l2/requirements.txt
aws-cdk-lib==2.173.3
constructs>=10.0.0,<11.0.0
boto3>=1.26
